Memory manager, builtin types
NEW: memory management has its own header. New Values are now marked in
the Scope it was created, so that popping a Scope releases Values
created in it (if there is no Scope below that is using the Value).

NEW: one may register types, this will be used primarily by builtin types.